MAT-4643: API so that a Measure can be shared with a user
MAT-4643: formatting
