+++ pom.xml
@@ -121,7 +121,7 @@
     <dependency>
       <groupId>gov.cms.madie</groupId>
       <artifactId>madie-java-models</artifactId>
-      <version>0.2.11-SNAPSHOT</version>
+      <version>0.2.12-SNAPSHOT</version>
     </dependency>
 
     <dependency>
